pyproject.toml
[tool.poetry]
name = "JCN-exchaige-assistant"
version = "0.1.0"
description = ""
authors = ["Jay Hsia <jayhsia1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
uvicorn = "*"
pydantic = "*"
pydantic-settings = "*"
python-dotenv = "*"
python-telegram-bot = "^20.6"
sentry-sdk = "*"
redis = "^5.0.1"
dependency-injector = "*"
bcrypt = "^4.1.2"
pytz = "*"
fastapi-limiter = "^0.1.6"
asyncpg = "^0.29.0"
sqlalchemy = "<2.0"
psycopg2-binary = "^2.9.9"
ujson = "^5.9.0"
python-multipart = "*"
pyotp = "^2.9.0"
pyjwt = "*"


[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
alembic = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
